requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.11
idna==3.7
python-dotenv==1.0.1
requests==2.31.0
soupsieve==2.5
sqlparse==0.5.0
typing_extensions==4.11.0
urllib3==2.2.1

telebot~=0.0.5